+++ repositories/withdrawal/base_repository.go
@@ -2,5 +2,5 @@ package withdrawal
 
 //BaseRepository defines basic methods
 type BaseRepository interface {
-	GetAvailableBanknotesValues() []int16
+	GetAvailableBanknotesValues() []int
 }
