+++ repositories/withdrawal/default_repository.go
@@ -3,8 +3,8 @@ package withdrawal
 type DefaultRepository struct {
 }
 
-func (d DefaultRepository) GetAvailableBanknotesValues() []int16 {
-	return []int16{50, 10, 5, 1}
+func (d DefaultRepository) GetAvailableBanknotesValues() []int {
+	return []int{50, 10, 5, 1}
 }
 
 func CreateDefaultRepository() BaseRepository {
